// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/forceTaskSuccessParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程
 * OpenAPI spec version: v1.0.0
 */

export type ForceTaskSuccessParams = {
  /**
   * No comments found.
   */
  workflowCode: string;
  /**
   * No comments found.
   */
  workflowVersionNo: number;
};