// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/listTaskInstancesParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程
 * OpenAPI spec version: v1.0.0
 */

export type ListTaskInstancesParams = {
  /**
   * No comments found.
   */
  workflowCode?: string;
  /**
   * No comments found.
   */
  workflowVersionNo?: number;
  /**
   * No comments found.
   */
  taskName?: string;
  /**
   * No comments found.
   */
  workflowInstanceName?: string;
  /**
   * No comments found.
   */
  status?: string;
  /**
   * No comments found.
   */
  startTimeFrom?: string;
  /**
   * No comments found.
   */
  endTimeTo?: string;
  /**
   * No comments found.
   */
  pageIndex?: number;
  /**
   * No comments found.
   */
  pageSize?: number;
};